chore(eventhub): organize test configurations
