+++ eventhub/wrangler.jsonc
@@ -5,23 +5,19 @@
 {
 	"$schema": "node_modules/wrangler/config-schema.json",
 	"name": "eventhub",
-	"main": "src/index.ts",
+	"main": "src/test.ts",
 	"compatibility_date": "2026-05-02",
 	"migrations": [
 		{
-			"new_sqlite_classes": ["EventHub", "TestEventHub"],
+			"new_sqlite_classes": ["TestEventHub"],
 			"tag": "v1",
 		},
 	],
 	"durable_objects": {
 		"bindings": [
 			{
-				"class_name": "EventHub",
-				"name": "EVENT_HUB",
-			},
-			{
 				"class_name": "TestEventHub",
-				"name": "EVENT_HUB_TEST",
+				"name": "EVENT_HUB",
 			},
 		],
 	},
